all functions finished for dashboard_2

+++ Backend/controller.js
@@ -2,11 +2,11 @@
 //Access-Control-Allow-Origin: any;
 const svr = require('./services')
 
-const login =  async (req, res) => {
+const login = async (req, res) => {
     const { username, password } = req.body;
     const rst = await svr.loginUser(username, password);
-    if(rst > 0) {
-        req.session = req.session || '' ;
+    if (rst > 0) {
+        req.session = req.session || '';
         req.session.user = rst;
         res.status(200);
         res.redirect('/dashboard');
@@ -16,10 +16,10 @@ const login =  async (req, res) => {
     }
 }
 
-const changePassword =  async (req, res) => {
+const changePassword = async (req, res) => {
     const { userId, newPassword } = req.body;
     const rst = await svr.changePassword(userId, newPassword);
-    if(rst){
+    if (rst) {
         res.status(200);
         res.send('Password Changed Successfully');
     } else {
@@ -30,33 +30,39 @@ const changePassword =  async (req, res) => {
 }
 
 const addTask = (req, res) => {
-    const userId = req.session.user;    
-    const {name, startTime, priority, category, reminderTime} = req.body;
+    const userId = req.session.user;
+    const { name, startTime, priority, category, reminderTime } = req.body;
     svr.addTask(name, startTime, priority, category, reminderTime, userId);
     res.send('Success')
     // Send appropriate response to the client
 }
 
-const showTaskByDate = (req, res) => {
-    const { date } = req.params;
-    svr.showTaskByDate(date);
-    // Send appropriate response to the client
+const showTaskByDate = async (req, res) => {
+    const date = req.query.date;
+    try {
+        const rst = await svr.showTaskByDate(date);
+        res.send(rst);
+    } catch (err) {
+        res.status(500); res.send(err);
+    }
 }
 
-const deleteTask =  async (req, res) => {
+const deleteTask = async (req, res) => {
     const { taskId } = req.body;
+    let checkUser = await svr.checkTaskUser(taskId, req.session.user);
+    if (!checkUser) { res.status(403); res.send(`403 Forbidden`); return; }
     const rst = await svr.deleteTask(taskId);
-    if(rst){
+    if (rst) {
         res.status(200);
         res.send(`Deleted task ${taskId} successfully.`)
     }
-    else{
+    else {
         res.status(404);
         res.send(`Could not delete task ${taskId}.`)
     }
 }
 
-const reminders =  (req, res) => {
+const reminders = (req, res) => {
     const { taskId } = req.params;
     const { reminderTime } = req.body;
     scheduleReminder(taskId, reminderTime);
@@ -70,19 +76,20 @@ const sendReminder = (req, res) => {
     // Send appropriate response to the client
 }
 
-async function queryTasks(req, res){
-    console.log('1');
+async function queryTasks(req, res) {
     const userId = req.session.user;
     const result = await svr.queryTasks(userId);
-    console.log(JSON.stringify(result));
+    //console.log(JSON.stringify(result));
     res.send(result);
 }
 
 
-const register = (req, res) => {
-    const {username, password} = req.body;
-    svr.addUser(username, password);
-    res.send('Success')
+const register = async (req, res) => {
+    const { username, password } = req.body;
+    try {
+        await svr.addUser(username, password);
+        res.redirect('/login');
+    } catch (err) { res.status(500); res.send(err); }
 }
 
 
@@ -95,5 +102,6 @@ module.exports = {
     deleteTask,
     reminders,
     queryTasks,
-    sendReminder
+    sendReminder,
+    register
 }
